Added adder example

+++ generics/src/main.rs
@@ -27,7 +27,7 @@ impl<T> Point<T>
 
 impl Point<f64>
 {
-    fn distance_from_origin(&self) -> f32
+    fn distance_from_origin(&self) -> f64
     {
         (self.x.powi(2) + self.y.powi(2)).sqrt()
     }
